Break apart the wallet list row styles
