+++ src/components/common/CryptoExchangeWalletListRow.js
@@ -1,16 +1,18 @@
 // @flow
+
 import { bns } from 'biggystring'
 import type { EdgeDenomination } from 'edge-core-js'
 import _ from 'lodash'
 import React, { Component } from 'react'
-import { Image, TouchableHighlight, View } from 'react-native'
+import { Image, StyleSheet, TouchableHighlight, View } from 'react-native'
 
 import * as intl from '../../locales/intl.js'
 import s from '../../locales/strings.js'
 import FormattedText from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
 import { calculateWalletFiatBalanceWithoutState } from '../../modules/UI/selectors.js'
-import { CryptoExchangeWalletListRowStyle as styles } from '../../styles/components/CryptoExchangeWalletListRowStyle.js'
+import { THEME } from '../../theme/variables/airbitz.js'
 import type { CustomTokenInfo, GuiWallet } from '../../types/types.js'
+import { scale } from '../../util/scaling.js'
 import { decimalOrZero, getFiatSymbol, truncateDecimals } from '../../util/utils.js'
 import { type TokenSelectObject, CryptoExchangeWalletListTokenRow } from './CryptoExchangeWalletListTokenRow.js'
 
@@ -48,7 +50,7 @@ type Props = StateProps & OwnProps
 
 const DIVIDE_PRECISION = 18
 
-class CryptoExchangeWalletListRow extends Component<Props, LocalState> {
+export class CryptoExchangeWalletListRow extends Component<Props, LocalState> {
   constructor(props: Props) {
     super(props)
     this.state = {
@@ -203,7 +205,7 @@ class CryptoExchangeWalletListRow extends Component<Props, LocalState> {
       return null
     }
     return (
-      <TouchableHighlight style={styles.touchable} underlayColor={styles.underlayColor} onPress={this.onPress}>
+      <TouchableHighlight underlayColor={THEME.COLORS.TRANSPARENT} onPress={this.onPress}>
         <View style={styles.rowContainerTop}>
           <View style={styles.containerLeft}>
             <Image style={styles.imageContainer} source={{ uri: wallet.symbolImage }} resizeMode="contain" />
@@ -251,4 +253,78 @@ class CryptoExchangeWalletListRow extends Component<Props, LocalState> {
   }
 }
 
-export { CryptoExchangeWalletListRow }
+const rawStyles = {
+  container: {
+    width: '100%'
+  },
+  rowContainerTop: {
+    width: '100%',
+    height: scale(76),
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingLeft: scale(10),
+    paddingRight: scale(10),
+    borderBottomWidth: scale(1),
+    borderBottomColor: THEME.COLORS.GRAY_3
+  },
+  rowContainerBottom: {
+    width: '100%'
+  },
+  containerLeft: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginRight: scale(10),
+    width: scale(36)
+  },
+  imageContainer: {
+    height: scale(24),
+    width: scale(24)
+  },
+  walletDetailsContainer: {
+    flex: 1,
+    flexDirection: 'column'
+  },
+  walletDetailsRow: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center'
+  },
+  walletDetailsRowCurrency: {
+    flex: 1,
+    fontSize: scale(18)
+  },
+  walletDetailsRowValue: {
+    textAlign: 'right',
+    fontSize: scale(18),
+    marginRight: scale(8),
+    color: THEME.COLORS.GRAY_1
+  },
+  walletDetailsRowName: {
+    flex: 1,
+    fontSize: scale(14),
+    color: THEME.COLORS.SECONDARY
+  },
+  walletDetailsRowFiat: {
+    fontSize: scale(14),
+    textAlign: 'right',
+    marginRight: scale(8),
+    color: THEME.COLORS.SECONDARY
+  },
+  walletHeaderContainer: {
+    backgroundColor: THEME.COLORS.GRAY_4,
+    flex: 3,
+    padding: scale(3),
+    paddingLeft: scale(15),
+    flexDirection: 'row',
+    paddingRight: scale(24)
+  },
+  walletHeaderTextContainer: {
+    flex: 1
+  },
+  walletHeaderText: {
+    color: THEME.COLORS.GRAY_2,
+    fontSize: scale(14)
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
